config-reader testcase
